refactor: Move post-analysis actions to a dedicated service and clean up action-runner logic
