+++ src/services/index.ts
@@ -5,3 +5,4 @@
 export * from './acrolinx-service.js'
 export * from './github-service.js'
 export * from './pr-comment-service.js'
+export * from './post-analysis-service.js'
